fix: conditional dart:io export for Web/WASM compatibility
Replaces the unconditional  with a
conditional re-export pattern: the real implementation is renamed to
, and a new
re-exports either that (when  is available) or a stub
() on Web/WASM.

Closes #163.
